Fix eslint errors, servin is boos, added lib folder for our node modules, add promise sql with pool connection
