--- scheepvaart-b00bs-firebase-adminsdk-z0nvx-aa2e8ae3e0.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "scheepvaart-b00bs",
-  "private_key_id": "aa2e8ae3e002e1aa2dbb1c93d5b77383c44a726e",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCpxFD/R7ssLTW8\ndl9KdV6YpkSE369rGIuPxPyaaXsemDWuXpXiV0Q51yN9VkLUn0q8JFPqYUrPW4cQ\n5JCRV0qqi8qdh2XKDWyjYeBZWVQVnB3jcgIxphZjgy4fJ7YlBFJk7dRTIZ9thxeK\nQcMOg7CXCq/9JCrBle28xfJKp8KQ24oJstRQqG52XWKeSmCzcksJdU4YLXUpmL2I\nbHnSILg0ov1XFUNd9phGX2gNeDqsafKOM8tDBmSiQ3sEP5/ZUOZhYmHzTZrVwYWU\n5yeYHMU20uyi8Epl3i88CDA5sQhWnUQrlQ4rq8U6xHyZ16pScwIY/lq4qEz7gI6t\nPW4S43DtAgMBAAECggEANOr9Zn2pEf2rtOnfDToluQaG8DY7m3IEutVn+gfsOV6v\nyQExzmJuYWs25FBZt6ZJqPjX8YTPSXCZ2wGsMKnZm+zERTNZ3tAtb4ZIZQGLzMqj\npGy601gWvYE8BLkftU0yE8TwvNMLu5vzEgUqtKEHECXkMs+txjciaof3XR8K7pUG\nYUH4SGYC1KuXghwC8e4f9p61hADyLqh+GrUW0A0Xq7eBgZToZBDGJJeflNG5XsJs\nO4YhSnXIXx+y7RTt4JIiu+FgCTEVNr0ZS3AqXty082hh1W9bEJ/qanG6jrgnwr+i\nk/p1IWEcf53Ylfgf0WKgdkjBMUXjRKdyw/mkKYX/DwKBgQDV4p0Y8KzGL9sfefxd\nfEIBBJeLS3RBxlzIworkSdJs1Nm6ca6QsMmotZWVuPqleiQDw+OnA21NZz+86q0C\nVgYHfE1xNPBEZDMKmrxuktJZrrnOiyRezQDy1KofsjFQbUDR7P9MEqM6SMCrZwua\nj0X0n2Ec2eD0QjSdCr7RCXq46wKBgQDLMdBRwtC9g5nWjCczqCjKVRWmWB/c4ZzA\npUjH5q6Nbh65MuDZZjBPeI4rX5iN9BAlyrDTFKhpofXJwcdcZizindOFTC6nF6Uz\nGJM5tNf8D/kfH3HWaBXljLQJd+togCXZQIh/nSgGAIDXb2c0URPHSx6eB8sXzcU7\nK62fqPiHhwKBgBNC+z7ZvdWDTcqB49ZPcGpDDZq/ytJ5mLWDv5pbg5XFfW4q+ruR\nEDgfUj9Pj1MDsuab546M4nUMDVNoFviU4qod6aN+Tvz5wRb9ZGbY1hS5hYYyRa47\n7A/yUvfTxlkl3aaKH1usegKvtpo5B3R93qDvEYHVpUqaA774zsJohiLNAoGBAMPm\n/fKmAfyFMb+RFoaOPO1+BMaAtg+OEUQG1tmsBWRj5AkoDXJsPG2STHdyzjEmOp9F\nASAxDYHyXXcsdZoXBHVkNreQUHZWrLZLnj4Hpijl+YhRF77gUSQkeLucO4Ff8522\nwyNOEHVg7NdIF9dvx8LgIN+5ocwMXniCzv1cumhnAoGAVgmfUMyu1pOAcDKJdZKR\ndyAyLTxn4CDqx6VsGeQfZ/VuStnivpC7UPtI6sSCIf5Xkj46un/Y3v4TvXiME2HA\n3bQmveQQhKxIYDHmCwb+2uocwbbVWk/Zj8hiT/YMsIp06bwbsIp9Nn6HqHikkh57\nZNeBC13lMmuXvXqX3Z472vI=\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-z0nvx@scheepvaart-b00bs.iam.gserviceaccount.com",
-  "client_id": "114625923451022113661",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://accounts.google.com/o/oauth2/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-z0nvx%40scheepvaart-b00bs.iam.gserviceaccount.com"
-}
